fix: Remove imports from conftest.py files, simplify the test files
