+++ tests/mongo_tests/conftest.py
@@ -1,79 +1,12 @@
-"""Shared Mongo test configuration and helpers."""
+"""Shared Mongo test fixtures."""
 
-import platform
-import sys
-from contextlib import suppress
-from urllib.parse import quote_plus
+import pytest
 
-from birch import Birch  # type: ignore[import-not-found]
-
-try:
-    from pymongo.mongo_client import MongoClient
-except (ImportError, ModuleNotFoundError):
-
-    class MongoClient:
-        """Mock MongoClient class raising ImportError on missing pymongo."""
-
-        def __init__(self, *args, **kwargs):
-            """Initialize the mock MongoClient."""
-            raise ImportError("pymongo is not installed!")
+from .helpers import _cleanup_mongo_client
 
 
-try:
-    from pymongo_inmemory import MongoClient as InMemoryMongoClient
-except (ImportError, ModuleNotFoundError):
-
-    class InMemoryMongoClient:
-        """Mock InMemoryMongoClient class.
-
-        Raises an ImportError on missing pymongo_inmemory.
-
-        """
-
-        def __init__(self, *args, **kwargs):
-            """Initialize the mock InMemoryMongoClient."""
-            raise ImportError("pymongo_inmemory is not installed!")
-
-
-class CfgKey:
-    HOST = "TEST_HOST"
-    PORT = "TEST_PORT"
-    TEST_VS_DOCKERIZED_MONGO = "TEST_VS_DOCKERIZED_MONGO"
-
-
-CFG = Birch(namespace="cachier", defaults={CfgKey.TEST_VS_DOCKERIZED_MONGO: False})
-_COLLECTION_NAME = f"cachier_test_{platform.system()}_{'.'.join(map(str, sys.version_info[:3]))}"
-
-
-def _get_cachier_db_mongo_client():
-    host = quote_plus(CFG[CfgKey.HOST])
-    port = quote_plus(CFG[CfgKey.PORT])
-    uri = f"mongodb://{host}:{port}?retrywrites=true&w=majority"
-    return MongoClient(uri)
-
-
-def _test_mongetter():
-    if not hasattr(_test_mongetter, "client"):
-        if str(CFG.mget(CfgKey.TEST_VS_DOCKERIZED_MONGO)).lower() == "true":
-            print("Using live MongoDB instance for testing.")
-            _test_mongetter.client = _get_cachier_db_mongo_client()
-        else:
-            print("Using in-memory MongoDB instance for testing.")
-            _test_mongetter.client = InMemoryMongoClient()
-    db_obj = _test_mongetter.client["cachier_test"]
-    if _COLLECTION_NAME not in db_obj.list_collection_names():
-        db_obj.create_collection(_COLLECTION_NAME)
-    return db_obj[_COLLECTION_NAME]
-
-
-def _cleanup_mongo_client():
-    """Close any cached Mongo test client safely."""
-    client = getattr(_test_mongetter, "client", None)
-    if client is None:
-        return
-    with suppress(Exception):
-        client._mongod._client.close()  # type: ignore[attr-defined]
-    with suppress(Exception):
-        client.close()
-    with suppress(Exception):
-        delattr(_test_mongetter, "client")
+@pytest.fixture(scope="session", autouse=True)
+def cleanup_mongo_client_fixture():
+    """Release cached Mongo client resources after the test session."""
+    yield
+    _cleanup_mongo_client()
