+++ tests/redis_tests/conftest.py
@@ -1,59 +1,29 @@
-"""Shared Redis test configuration and helpers."""
+"""Shared Redis test fixtures."""
+
+from contextlib import suppress
 
 import pytest
-from birch import Birch  # type: ignore[import-not-found]
 
-from .clients import _MockRedis
-
-try:
-    import redis
-
-    REDIS_AVAILABLE = True
-except ImportError:
-    redis = None  # type: ignore[assignment]
-    REDIS_AVAILABLE = False
+from .clients import _AsyncInMemoryRedis, _MockRedis
+from .helpers import _get_test_redis_client
 
 
-class CfgKey:
-    HOST = "TEST_REDIS_HOST"
-    PORT = "TEST_REDIS_PORT"
-    DB = "TEST_REDIS_DB"
-    TEST_VS_DOCKERIZED_REDIS = "TEST_VS_DOCKERIZED_REDIS"
+@pytest.fixture
+def sync_redis_client_fixture():
+    """Yield a fresh sync Redis client for each test, cleared after use."""
+    live_client = _get_test_redis_client()
+    client = _MockRedis() if live_client is None else live_client
+    yield client
+    with suppress(Exception):
+        if hasattr(client, "data"):
+            client.data.clear()
+        else:
+            client.flushdb()
 
 
-CFG = Birch(
-    namespace="cachier",
-    defaults={CfgKey.TEST_VS_DOCKERIZED_REDIS: False},
-)
-
-
-def _get_test_redis_client():
-    """Get a Redis client for testing."""
-    if not REDIS_AVAILABLE:
-        pytest.skip("Redis not available")
-
-    if str(CFG.mget(CfgKey.TEST_VS_DOCKERIZED_REDIS)).lower() == "true":
-        print("Using live Redis instance for testing.")
-        host = CFG.get(CfgKey.HOST, "localhost")
-        port = int(CFG.get(CfgKey.PORT, 6379))
-        db = int(CFG.get(CfgKey.DB, 0))
-        try:
-            client = redis.Redis(host=host, port=port, db=db, decode_responses=False)
-            client.ping()
-            return client
-        except redis.ConnectionError as e:
-            print(f"Failed to connect to Redis: {e}")
-            pytest.skip("Redis not available")
-    else:
-        print("Using mock Redis for testing.")
-        return None
-
-
-def _test_redis_getter():
-    """Get Redis client for testing."""
-    client = _get_test_redis_client()
-    if client is None:
-        if not hasattr(_test_redis_getter, "_mock_client"):
-            _test_redis_getter._mock_client = _MockRedis()
-        return _test_redis_getter._mock_client
-    return client
+@pytest.fixture
+def async_redis_client_fixture():
+    """Yield a fresh async in-memory Redis client for each test, cleared after use."""
+    client = _AsyncInMemoryRedis()
+    yield client
+    client._data.clear()
